Replace static image of the osm-liberty style with interactive map
This change removes the static image `img/osm-liberty.jpg` from the styles page.
Instead it uses the style located at http://osm-liberty.lukasmartinelli.ch/style.json to render an interactive map like for the other styles on the page.
The zoom level is limited to >=z6, because the Natural Earth tiles used for lower zoom levels are not served over HTTPS.